delay update of receivedItemIndex to event engine
This should hopefully fix the bug with items being dropped upon
reloading a save.

Previously, the client would increment `receivedItemIndex` when writing
the new item into the game's memory, a pattern that was copied from
the Emerald client. However, this could lead to a race in which the
index is incremented *and the game is saved* before the item is actually
received, leading to the item being dropped if the game is then
reloaded. The fix is to instead only increment `receivedItemIndex` after
the game has actually acknowledged receipt of the remote item (by
displaying a pop-up, etc).

Signed-off-by: Cameron Wong <cam@camdar.io>